requirements.txt
xrpl-py==4.3.0
python-dotenv==1.2.1
requests==2.32.5
